+++ tests/deprecated_api/test_remove_1-6.py
@@ -15,6 +15,7 @@
 import pytest
 
 from pytorch_lightning import Trainer
+from pytorch_lightning.callbacks.early_stopping import EarlyStopping
 from pytorch_lightning.plugins.training_type import DDPPlugin, DDPSpawnPlugin
 from tests.helpers import BoringDataModule, BoringModel
 
@@ -172,3 +173,11 @@ def test_v1_6_0_datamodule_hooks_calls(tmpdir):
     assert dm.prepare_data_calls == 1
     assert dm.setup_calls == ['fit', None]
     assert dm.teardown_calls == ['validate', 'test']
+
+
+def test_v1_6_0_early_stopping_monitor(tmpdir):
+    with pytest.deprecated_call(
+        match=r"The `EarlyStopping\(monitor\)` argument will be required starting in v1.6."
+        " For backward compatibility, setting this to `early_stop_on`."
+    ):
+        EarlyStopping()